server/requirements.txt
flask
firebase-admin
python-dotenv
transformers
torch
openai
requests
flask-cors
gunicorn